requirements.txt
python-kafka
# numpy for testing
